+++ otx/algorithms/common/tasks/base_task.py
@@ -19,9 +19,11 @@ import os
 import shutil
 import tempfile
 from abc import ABC, abstractmethod
+from datetime import timedelta
 from typing import Any, Dict, Iterable, List, Optional
 
 import torch
+from torch import distributed as dist
 
 from otx.algorithms.common.adapters.mmcv.hooks import OTXLoggerHook
 from otx.algorithms.common.adapters.mmcv.hooks.cancel_hook import CancelInterfaceHook
@@ -97,11 +99,7 @@ class OTXTask(IInferenceTask, IExportTask, IEvaluationTask, IUnload, ABC):
         self._labels = task_environment.get_labels(include_empty=False)
         self._work_dir_is_temp = False
         self._output_path = output_path
-        if self._output_path is None:
-            self._output_path = tempfile.mkdtemp(prefix="OTX-task-")
-            self._work_dir_is_temp = True
-        else:
-            os.makedirs(self._output_path, exist_ok=True)
+        self._output_path = output_path if output_path is not None else self._get_tmp_dir()
         self._time_monitor: Optional[TimeMonitorCallback] = None
         self.on_hook_initialized = OnHookInitialized(self)
         self._learning_curves = UncopiableDefaultDict(OTXLoggerHook.Curve)
@@ -119,6 +117,32 @@ class OTXTask(IInferenceTask, IExportTask, IEvaluationTask, IUnload, ABC):
 
         # This is for hpo, and this should be removed
         self.project_path = self._output_path
+
+        if self._is_multi_gpu_training():
+            self._setup_multigpu_training()
+
+    @staticmethod
+    def _is_multi_gpu_training():
+        multi_gpu_env = ["MASTER_ADDR", "MASTER_PORT", "LOCAL_WORLD_SIZE", "WORLD_SIZE", "LOCAL_RANK", "RANK"]
+        for env in multi_gpu_env:
+            if env not in os.environ:
+                return False
+
+        return torch.cuda.is_available()
+
+    @staticmethod
+    def _setup_multigpu_training():
+        if not dist.is_initialized():
+            torch.cuda.set_device(int(os.environ["LOCAL_RANK"]))
+            dist.init_process_group(backend="nccl", init_method="env://", timeout=timedelta(seconds=30))
+            logger.info(f"Dist info: rank {dist.get_rank()} / {dist.get_world_size()} world_size")
+
+    def _get_tmp_dir(self):
+        self._work_dir_is_temp = True
+        # If training is excuted with torchrun, set all trainings' output directory same
+        if "TORCHELASTIC_RUN_ID" in os.environ:
+            return os.path.join(tempfile.gettempdir(), f"OTX-task-torchelastic-{os.environ['TORCHELASTIC_RUN_ID']}")
+        return tempfile.mkdtemp(prefix="OTX-task-")
 
     def _load_model(self):
         """Loading model from checkpoint."""